Break `src/utils.h` into more cohesive smaller modules
