+++ src/command_lint.cc
@@ -12,7 +12,11 @@
 #include <sstream>  // std::ostringstream
 
 #include "command.h"
+#include "configuration.h"
 #include "error.h"
+#include "input.h"
+#include "logger.h"
+#include "resolver.h"
 #include "utils.h"
 
 template <typename Options, typename Iterator>
@@ -21,10 +25,10 @@ static auto disable_lint_rules(sourcemeta::core::SchemaTransformer &bundle,
                                Iterator last) -> void {
   for (auto iterator = first; iterator != last; ++iterator) {
     if (bundle.remove(std::string{*iterator})) {
-      sourcemeta::jsonschema::cli::log_verbose(options)
+      sourcemeta::jsonschema::LOG_VERBOSE(options)
           << "Disabling rule: " << *iterator << "\n";
     } else {
-      sourcemeta::jsonschema::cli::log_verbose(options)
+      sourcemeta::jsonschema::LOG_VERBOSE(options)
           << "warning: Cannot exclude unknown rule: " << *iterator << "\n";
     }
   }
@@ -46,10 +50,9 @@ static auto reindent(const std::string_view &value,
   }
 }
 
-static auto
-get_lint_callback(sourcemeta::core::JSON &errors_array,
-                  const sourcemeta::jsonschema::cli::InputJSON &entry,
-                  const bool output_json) -> auto {
+static auto get_lint_callback(sourcemeta::core::JSON &errors_array,
+                              const sourcemeta::jsonschema::InputJSON &entry,
+                              const bool output_json) -> auto {
   return [&entry, &errors_array,
           output_json](const auto &pointer, const auto &name,
                        const auto &message, const auto &result) {
@@ -103,7 +106,7 @@ get_lint_callback(sourcemeta::core::JSON &errors_array,
   };
 }
 
-auto sourcemeta::jsonschema::cli::lint(const sourcemeta::core::Options &options)
+auto sourcemeta::jsonschema::lint(const sourcemeta::core::Options &options)
     -> void {
   const bool output_json = options.contains("json");
 
@@ -134,7 +137,7 @@ auto sourcemeta::jsonschema::cli::lint(const sourcemeta::core::Options &options)
     }
 
     for (const auto &only : options.at("only")) {
-      log_verbose(options) << "Only enabling rule: " << only << "\n";
+      LOG_VERBOSE(options) << "Only enabling rule: " << only << "\n";
       if (blacklist.erase(only) == 0) {
         throw InvalidLintRuleError{"The following linting rule does not exist",
                                    std::string{only}};
@@ -189,7 +192,7 @@ auto sourcemeta::jsonschema::cli::lint(const sourcemeta::core::Options &options)
 
       const auto &custom_resolver{
           resolver(options, options.contains("http"), dialect, configuration)};
-      log_verbose(options) << "Linting: " << entry.first.string() << "\n";
+      LOG_VERBOSE(options) << "Linting: " << entry.first.string() << "\n";
       if (entry.first.extension() == ".yaml" ||
           entry.first.extension() == ".yml") {
         throw YAMLInputError{
@@ -241,7 +244,7 @@ auto sourcemeta::jsonschema::cli::lint(const sourcemeta::core::Options &options)
       const auto dialect{default_dialect(options, configuration)};
       const auto &custom_resolver{
           resolver(options, options.contains("http"), dialect, configuration)};
-      log_verbose(options) << "Linting: " << entry.first.string() << "\n";
+      LOG_VERBOSE(options) << "Linting: " << entry.first.string() << "\n";
 
       const auto wrapper_result = sourcemeta::jsonschema::try_catch([&]() {
         try {
